Fixed the bug in sonar in the pokeballDrawer.

+++ src/main/java/cpaintus/controllers/drawers/PokeballDrawerStrategy.java
@@ -25,7 +25,7 @@ public class PokeballDrawerStrategy implements IDrawerStrategy {
 		double width = shape.getWidth();
 		double height = shape.getHeight();
 		
-		int lineWidth = shape.getLineWidth();
+		double lineWidth = shape.getLineWidth();
 		
 		// Colored top of the ball
 		gc.fillArc(lineWidth/2,
